# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation untuk Frontend",
        "title": "Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/angkatan-mapala": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Admin mengambil semua data angkatan mapala",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "angkatan-mapala"
                ],
                "summary": "Get all angkatan mapala",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Admin membuat angkatan mapala baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "angkatan-mapala"
                ],
                "summary": "Create angkatan mapala",
                "parameters": [
                    {
                        "description": "Angkatan mapala payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/angkatan.AngkatanMapalaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/admin/angkatan-mapala/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Admin update data angkatan mapala",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "angkatan-mapala"
                ],
                "summary": "Update angkatan mapala",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Angkatan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/angkatan.AngkatanMapalaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Admin menghapus angkatan mapala",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "angkatan-mapala"
                ],
                "summary": "Delete angkatan mapala",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Angkatan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/auth/admin-reset": {
            "post": {
                "description": "Admin mengajukan permintaan reset password berdasarkan identitas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-auth"
                ],
                "summary": "Request admin password reset",
                "parameters": [
                    {
                        "description": "Forgot password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permintaan reset berhasil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "angkatan.AngkatanMapalaRequest": {
            "type": "object",
            "properties": {
                "nama": {
                    "type": "string",
                    "example": "Mapala 2023"
                }
            }
        },
...
        "request.ForgotPasswordRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "nra": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT token tanpa prefix Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}